host 비로그인시 다른페이지 접근불가 추가

+++ src/hostController/roomDeleteServlet.java
@@ -6,6 +6,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import model.StudioResDAO;
 
@@ -17,10 +18,12 @@ public class roomDeleteServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
        
 
-	/**
-	 * @see HttpServlet#doGet(HttpServletRequest request, HttpServletResponse response)
-	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		HttpSession session = request.getSession();
+		if(session.getAttribute("host_no") == null) {
+			response.sendRedirect("hostLogin");
+			return;
+		}
 		StudioResDAO dao = new StudioResDAO();
 		int room_no = Integer.parseInt(request.getParameter("room_no"));
 		dao.deleteRoom(room_no);
